test: move collection endpoint tests to separate files
